pyproject.toml
[tool.poetry]
name = "test_for_mobicult"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.5.1"
requests = "^2.32.4"
django-stubs = "^5.2.2"
mypy-extensions = "^1.1.0"
types-requests = "^2.32.4.20250611"

env = ["USE_LOCAL_TEST_CONFIG=true"]
asyncio_mode = 'auto'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"